% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picalc.R
\name{filter.hq.genes}
\alias{filter.hq.genes}
\title{Filter high quality genes}
\usage{
filter.hq.genes(gff.file, peptide.fasta, bed = F, gene.identifier = "Parent")
}
\arguments{
\item{gff.file}{Path to the input gff3 file with gene models.}

\item{peptide.fasta}{Path to the input fasta file with protein sequences of the gene models.}

\item{bed}{Logical, whether to return coordinates in a bed format (True; 0-based) or gff3 coordinates (False, default; starts from 1).}

\item{gene.identifier}{The unique identifier for matching exons with gene models in the gff3 file. Default is "Parent"}
}
\value{
A list with two items: HQ - high quality gene models and ALL - all gene models. Each list item is a reduced gff3 file.
}
\description{
This function selects high quality genes from the set of annotated genes in a given gff3 file.
It checks that the gene length is divisible by three (so that codons are intact) and that the
translated protein starts with a methionine.
}
