% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picalc.R
\name{folds.0}
\alias{folds.0}
\title{Identify zero-fold degenerate positions (ie non-synonymous sites) from the gene models.}
\usage{
folds.0(gff)
}
\arguments{
\item{gff.file}{The input gff3 file with gene models. Recommended input: high quality gene models from filter.hq.genes}
}
\value{
A data frame containing coordinates of non-synonymous positions.
}
\description{
Identify zero-fold degenerate positions (ie non-synonymous sites) from the gene models.
}
