% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picalc.R
\name{folds.4}
\alias{folds.4}
\title{Identify four-fold degenerate positions (4dtv) from the gene models.}
\usage{
folds.4(gff, peptide.fasta)
}
\arguments{
\item{peptide.fasta}{Path to the input fasta file with protein sequences of the gene models.}

\item{gff.file}{The input gff3 file with gene models. Recommended input: high quality gene models from filter.hq.genes}
}
\value{
A data frame containing coordinates of 4dtv sites.
}
\description{
Identify four-fold degenerate positions (4dtv) from the gene models.
}
