% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picalc.R
\name{Pi_N.par}
\alias{Pi_N.par}
\title{Calculate Pi_N for loci with a high impact or moderate impact mutation}
\usage{
Pi_N.par(TH, exon.sel, exon.mut, nSNP = NA, cores = 15)
}
\description{
Calculate Pi_N for loci with a high impact or moderate impact mutation
}
