% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picalc.R
\name{Pi_S.par}
\alias{Pi_S.par}
\title{Calculate Pi_S based on intergenic loci.
This is obtained by excluding all exonic loci.}
\usage{
Pi_S.par(TH, exon.all, cores = 15)
}
\description{
Calculate Pi_S based on intergenic loci.
This is obtained by excluding all exonic loci.
}
